release 1.0.0

+++ RJPodDemo/Classes/EngineDemo.h
@@ -12,6 +12,6 @@
 + (NSInteger)sumOfValue:(NSInteger)firstValue SecondValue:(NSInteger)secondValue;
 + (CGFloat)subtractValueFrom:(CGFloat)firstValue toSecondValue:(CGFloat)secondValue;
 + (CGFloat)multiplyValueFrom:(CGFloat)firstValue toSecondValue:(CGFloat)secondValue;
-+ (CGFloat)sqrtValue:(NSInteger)value;
++ (CGFloat)sqrtValue:(CGFloat)value;
 
 @end
